requirements.txt
torch
torchvision
facenet-pytorch
opencv-python
numpy
Pillow